Extract ffi into separate crate
