+++ rust/android_build_all.sh
@@ -1,3 +1,3 @@
-cargo build --target i686-linux-android
-cargo build --target armv7-linux-androideabi
-cargo build --target aarch64-linux-android
+cargo build -p toodle_ffi --target i686-linux-android
+cargo build -p toodle_ffi --target armv7-linux-androideabi
+cargo build -p toodle_ffi --target aarch64-linux-android
