+++ rust/src/utils.rs
@@ -8,7 +8,6 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
-extern crate ffi_utils;
 extern crate time;
 
 extern crate mentat;
