+++ rust/toodle_ffi/src/ctypes.rs
@@ -11,11 +11,6 @@
 use std::os::raw::c_char;
 use std::ptr;
 
-use ffi_utils::strings::{
-    string_to_c_char,
-    c_char_to_string,
-};
-
 use time::{
     Timespec,
 };
@@ -24,9 +19,14 @@ use mentat::{
     Uuid,
 };
 
-use items::{
+use toodle::items::{
     Item,
     Items,
+};
+
+use utils::strings::{
+    string_to_c_char,
+    c_char_to_string,
 };
 
 #[repr(C)]
